+++ CHANGELOG.md
@@ -5,6 +5,11 @@ This file uses change log convention from [Keep a CHANGELOG].
 
 ## [Unreleased]
 
+## [1.3.0] - 2020-12-30
+### Fixed
+- Fixed resolving paths with different working directory than `tsconfig.json` (Use the new `cwd` option)
+- Fixed support for multiple imports per line
+
 ## [1.2.0] - 2020-12-29
 ### Fixed
 - Fixed paths inside comments being resolved.
@@ -60,7 +65,8 @@ This file uses change log convention from [Keep a CHANGELOG].
 [Keep a CHANGELOG]: http://keepachangelog.com
 [Semantic Versioning]: http://semver.org/
 
-[unreleased]: https://github.com/dhkatz/gulp-ts-alias/compare/1.2.0...HEAD
+[unreleased]: https://github.com/dhkatz/gulp-ts-alias/compare/1.3.0...HEAD
+[1.3.0]: https://github.com/dhkatz/gulp-ts-alias/compare/1.2.0...1.3.0
 [1.2.0]: https://github.com/dhkatz/gulp-ts-alias/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/dhkatz/gulp-ts-alias/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/dhkatz/gulp-ts-alias/compare/0.2.1...1.0.0